+++ GUI/MainForm.cs
@@ -713,9 +713,9 @@ namespace GUI
             {
                 return new Types.Viewers.Audio().Create(vrfGuiContext, stream);
             }
-            else if (Types.Viewers.FaceExpressionData.IsAccepted(magic))
+            else if (Types.Viewers.FlexSceneFile.IsAccepted(magic))
             {
-                return new Types.Viewers.FaceExpressionData().Create(vrfGuiContext, stream);
+                return new Types.Viewers.FlexSceneFile().Create(vrfGuiContext, stream);
             }
 
             return new Types.Viewers.ByteViewer().Create(vrfGuiContext, stream);
